+++ config.env
@@ -4,7 +4,7 @@ NODE_ENV=development
 # DB_PASS=VUvCDJzAS2mmKqza
 # DB_NAME=ecommore-db
 # DB_USER=fahd
-DB_URI=mongodb://fahd:01140@ac-3nktzhy-shard-00-00.pc8fcir.mongodb.net:27017,ac-3nktzhy-shard-00-01.pc8fcir.mongodb.net:27017,ac-3nktzhy-shard-00-02.pc8fcir.mongodb.net:27017/?ssl=true&replicaSet=atlas-fn6ldv-shard-0&authSource=admin&retryWrites=true&w=majority
+DB_URL=mongodb://127.0.0.1:27017/e-commerce-api
 BASE_URL=http://localhost:8000
 JWT_SECRET_KEY=Fahd is Backend NodeJS Programmer
 JWT_EXPIRE_TIME=90d
