+++ config/datebase.js
@@ -1,11 +1,13 @@
-
-const mongoose=require('mongoose')
-const dbConnection=()=>{
-    mongoose.connect(process.env.DB_URI).then((conn)=>{
-        console.log(` Database Connected ${conn.connection.host}`);
-    }).catch((err)=>{
-        console.error(`Database Error:${err}`)
-        process.exit(1)
-    })
-}
-module.exports=dbConnection
+const mongoose = require("mongoose");
+const dbConnection = async () => {
+  try {
+    const conn = await mongoose.connect(
+      process.env.DB_URL || "mongodb://127.0.0.1:27017/e-commerce-api"
+    );
+    console.log(`✅ MongoDB Connected: ${conn.connection.host}`);
+  } catch (error) {
+    console.error(error);
+    process.exit(1);
+  }
+};
+module.exports = dbConnection;
