+++ routes/couponRoute.js
@@ -1,0 +1,16 @@
+const express = require("express");
+const Authservices = require("../services/authService");
+const {
+  getAllCopons,
+  getCopon,
+  createCopon,
+  updateCopon,
+  deleteCopon,
+} = require("../services/couponService");
+
+const router = express.Router();
+router.use(Authservices.protect, Authservices.allowTo("admin", "manager"));
+
+router.route("/").post(createCopon).get(getAllCopons);
+router.route("/:id").delete(deleteCopon).get(getCopon).put(updateCopon);
+module.exports = router;
