--- routes/wichListRoute.js
@@ -1,14 +1,0 @@
-const express = require("express");
-const Authservices = require("../services/authService");
-const {
-  addProductToWichlist,
-  removeProductFromWichlist,
-  userWishList,
-} = require("../services/wishListService");
-
-const router = express.Router();
-router.use(Authservices.protect, Authservices.allowTo("user"));
-
-router.route("/").post(addProductToWichlist).get(userWishList);
-router.route("/:productId").delete(removeProductFromWichlist);
-module.exports = router;
